frontend/requirements.txt
streamlit==1.22.0
python-dotenv
nomic
matplotlib
seaborn
cleanvision
